logic cleanup
